+++ spec/lib/cassie/tasks/task_runner_spec.rb
@@ -2,15 +2,12 @@ require 'cassie/tasks'
 
 RSpec.describe Cassie::Tasks::TaskRunner do
   let(:klass){ Cassie::Tasks::TaskRunner }
-  let(:object){ klass.new }
+  let(:object){ klass.new(args) }
   let(:args){ Array(command) + params }
   let(:params){ [] }
   let(:command){ nil }
 
-  describe "#run_command" do
-    def run
-      object.run_command(args)
-    end
+  describe "#run" do
     Rake.application.tasks.each do |t|
       context "with the #{t.name} task" do
         let(:command){ t.name.sub("cassie:", "") }
@@ -18,7 +15,7 @@ RSpec.describe Cassie::Tasks::TaskRunner do
         it "finds an associated command" do
           task = Rake::Task["cassie:#{command}"]
           expect(task).to receive(:invoke)
-          run
+          object.run
         end
       end
     end
@@ -50,7 +47,7 @@ RSpec.describe Cassie::Tasks::TaskRunner do
 
         it "runs an associated task" do
           expect(task).to receive(:invoke)
-          run
+          object.run
         end
       end
     end